Move vector functionality updates
This time, the vector module can be used in other projects, as it is completely standalone (relying only on python's standard library), due to it gaining the VectorOutOfBoundsError class. There are now methods that can get the angle between a vector and VECTOR_i. The functionality will be kept to only what the project needs though.